DQSHBE-7 Move ShellButtonItemImpl inner class to a separate file for better modularity
Update ShellButtonImplTest with proper indentation

Add Logs and Comments to the java classes

Update styles in scss
