+++ core/src/test/resources/shellbutton/ShellButtonImplTest.json
@@ -1,21 +1,17 @@
 {
     "shell-buttons": {
         "buttons": {
-            "btn-1":{
+            "btn-1": {
                 "buttonText": "Button 1",
                 "buttonLink": "shell/us/en/home-page-1",
                 "icon": "content/shell/images/icons/icon-1.png"
             },
-            "btn-2":{
+            "btn-2": {
                 "buttonText": "Button 2",
                 "buttonLink": "shell/us/en/home-page-2",
                 "icon": "content/shell/images/icons/icon-2.png"
             }
         }
     },
-"shell-buttonsEmpty": {
-       
-    }
-
-
+    "shell-buttonsEmpty": {}
 }
