+++ ui.apps/src/main/content/jcr_root/apps/shell/components/content/shell-button/.content.xml
@@ -6,8 +6,6 @@
     jcr:primaryType="cq:Component"
     jcr:title="Shell Button"
     componentGroup="Shell"
-    sling:resourceType="shell/components/content/shell-button"
-    cq:styleDefaultClasses="btn-default"
-    cq:styleGroups="[appearance]">
-    
+    container="false"
+    sling:resourceType="shell/components/content/shell-button">
 </jcr:root>
