+++ ui.apps/src/main/content/jcr_root/apps/shell/components/content/shell-button/_cq_design_dialog/.content.xml
@@ -12,7 +12,6 @@
         jcr:primaryType="nt:unstructured"
         sling:resourceType="granite/ui/components/coral/foundation/container">
         <items jcr:primaryType="nt:unstructured">
-
             <styletab
                 jcr:primaryType="nt:unstructured"
                 sling:resourceType="granite/ui/components/coral/foundation/include"
