+++ ui.apps/src/main/content/jcr_root/apps/shell/components/content/shell-button/_cq_dialog/.content.xml
@@ -10,7 +10,6 @@
         jcr:primaryType="nt:unstructured"
         sling:resourceType="granite/ui/components/coral/foundation/container">
         <items jcr:primaryType="nt:unstructured">
-
             <properties
                 jcr:primaryType="nt:unstructured"
                 jcr:title="Properties"
@@ -67,7 +66,6 @@
                     </columns>
                 </items>
             </properties>
-
         </items>
     </content>
 </jcr:root>
