+++ ui.apps/src/main/content/jcr_root/apps/shell/components/content/shell-button/shell-button.html
@@ -1,22 +1,19 @@
-<sly class="button-component"
-     data-sly-use.buttonModel="com.shell.core.models.ShellButton"
-     data-sly-use.placeholderTemplate="core/wcm/components/commons/v1/templates.html"
-     data-sly-test.hasContent="${!buttonModel.empty}">
-
-        <sly data-sly-list="${buttonModel.buttons}">
-            <li class="button-item">
-                <a href="${item.buttonLink @ context='uri'}" class="button-link">
-                    <sly data-sly-test="${item.icon}">
-                        <img src="${item.icon @ context='uri'}" alt="Icon" class="button-icon" />
-                    </sly>
-                    
-                    <span class="button-text">${item.buttonText}</span>
-                   
-                    </a>
-            </li>
+<sly data-sly-use.clientlib="core/wcm/components/commons/v1/templates/clientlib.html"/>
+<sly 
+  data-sly-use.buttonModel="com.shell.core.models.ShellButton"
+  data-sly-use.placeholderTemplate="core/wcm/components/commons/v1/templates.html"
+  data-sly-test.hasContent="${!buttonModel.empty}">
+  <div class="button-container">
+  <sly data-sly-list="${buttonModel.buttons}">
+    <li class="button-item">
+      <a href="${item.buttonLink}.html" class="button-link">
+        <sly data-sly-test="${item.icon}">
+          <img src="${item.icon}" alt="Icon" class="button-icon"/>
         </sly>
-
-   
-
-    <sly data-sly-call="${placeholderTemplate.placeholder @ isEmpty=!hasContent}"></sly>
+        <span class="button-text">${item.buttonText}</span>
+      </a>
+    </li>
+    </div>
+  </sly>
 </sly>
+<sly data-sly-call="${placeholderTemplate.placeholder @ isEmpty=!hasContent}"></sly>
